feat: Zc companion struct for instruction data deserialization
Extract shared Zc helpers (map_to_pod_type, zc_serialize_field,
zc_deserialize_field) from account.rs into helpers.rs. Use the same
pattern for instruction data: generate an InstructionDataZc struct
with alignment-1 Pod types and a compile-time assert, replacing
read_unaligned with a direct pointer cast.
